refactor: Export createUser function in userOperations.ts
Export the `createUser` function in `userOperations.ts` to make it accessible from other modules. This change allows other parts of the codebase to use the `createUser` function for creating new users with hashed passwords. The function now has a wider scope and can be imported and used in other files.